use guava cache

+++ CF-PredictorBackEnd/src/main/java/com/wslfinc/cf/sdk/CodeForcesSDK.java
@@ -177,7 +177,7 @@ public class CodeForcesSDK {
         RatingProcessing.removeFromCache(contestId);
     }
 
-    public static void changeTimeToUodate(long timeToUpdateMS) {
+    public static void changeTimeToUodate(int timeToUpdateMS) {
         RatingProcessing.changeTimeToUpdate(timeToUpdateMS);
     }
 }
